kb agent tests split
